Remove german special page alias for 'Captcha'
The translation is wrong, because it only contains the edit part of
Captcha, but there are also captchas for action creation, login
throttling and add url, which using the same special page.

translatewiki.net does not allow aliases, which are the same than the
english word. That means the field on Special:AdvancedTranslate is empty
and animated people to fill it up. On the other hand, removing aliases
is not supported.

Follow up Ie8fa19cf (same release, so removing should not be a problem)

Change-Id: I81c4b69923001f72f88587507e3599b529c75579