Get Asirra captcha working on more forms
The jQuery selector to find the form only worked for create account;
it now works for the captchas on login password failures and on page
editing.

Remove two debugging console.log() lines and fix some JSHint errors.

Rename the message for login captchas from the unused asirra-badpass to
the correct asirra-badlogin.

Bug: 39536
Change-Id: I92c3cb93337d4a704221d5a0e50ed56ae3b6b1cf