Position the "Refresh" link over the image
This changes the CSS of the .fancycaptcha-image-container to allow
positioning within it, and repositions .fancycaptcha-reload at its
lower-right.  The container is taller than nearly all images, but
there's a chance the ↻Refresh link might obscure part of a captcha
image.

This is compatible with the new Create account form's repositioning of
the FancyCaptcha inside a gray frame using JS, though gerrit I003c5f1b75
improves Create account once this is deployed.

Bug: 47699
Change-Id: I57be0bc8aa80b9ea6148db979ad28397d2107b86