Add more words to blacklist
In adition to profanities, remove silly or questionable words
thus reducing the possibility of captchas like "killkids" or
"allahlies". I understand that this list is still nowhere near
complete, but at least this is a step in the right direction.

For the reference, the current wordlist used for captcha generation
is http://noc.wikimedia.org/~maxsem/words.txt
(words with anything but lowercase letters also get discarded)

Change-Id: Ic1cb0aa688811886e013bc819866abd745d02f68