Move apihelp messages to a separate file
This is one of the last extensions to be converted to having
the API messages in a separate file (at least out of the extensions
that are used be Wikimedia).

This one is a bit different from the others because it actually as several
extensions with separate i18n files, so it requires extra-careful review,
especially in Gruntfile and the extension.json files.

Bug: T189982
Change-Id: I66faae6fd4ff447327587c89ad2a1704edd1b356