Add Turnstile support
This is a rudimentary implementation of Cloudflare's Turnstile captcha alternative.

Since it is fairly similar to ReCaptchaNoCaptcha, there is some duplicate code used here. I opted not to use the migration path that CF has for reCAPTCHA v2 and build this as a separate implementation. After T324925, it should be easy to refactor this code.

Users should set $wgTurnstileSiteKey and $wgTurnstileSecretKey as appropriate, based on details from the Cloudflare dashboard.

Bug: T319068
Change-Id: I0d5a74655619975f0bf61b5b672159afe5f4fb00