QuestyCaptcha: Remove mw-ui-input class that does nothing
It does nothing, because the 'mediawiki.ui.input' module is not loaded
on the edit form, so it doesn't apply any styles.

I'm not sure if it ever worked, but it definitely hasn't worked since
at least 2021 (that's the oldest wiki I found on Patchdemo that has
the class but doesn't have the styling).

Change-Id: Ibd9d3df3aa74eb22c6b7fed97076056fc4187c66