LoginAttemptCounter: Add extra hardening for long period too
The current method of showing captcha with 3 attempts in 5 minutes is
nice but it would be even better to show captcha in more attempts but
longer period of time, the default roughly translates to showing captcha
if bad login is more than 100 times in every day.

It is extremely ugly but it already has todo to migrate to Throttler,
once that's in place, it could be simplified a lot.

Change-Id: I8bfcd52f4c77946182ca7c58b31fd8b38bbcd3cf