Allow $wgCaptchaClass to be set to a built-in module by name
Discourage exposing internal PHP class names as public APIs in their
own right. For compat with additional extensions to ConfirmEdit, it
can also be set to any custom subclass like before.

This should improve ergonomics, as there wasn't a list of
fully-qualified class names in the README, whereas it's fairly simple
to make a list of symbolic stable names.

These are also compatible with the values that were used prior to
MW 1.42, and thus make it so that we can remove the class aliases
without breaking existing installations that set their LocalSettings
prior to that, and without going through a painful deprecation-for-
having-a-value-in-LocalSettings phase (T384064).

Bug: T384064
Change-Id: Ifa42da394e25a8d87ad8bcc2eec16c9efe91428c
(cherry picked from commit fb0423a1a80dfae5796f8be1b02068aec03df173)