hCaptcha: Use RL's simpler config mechanism
Rather than using addJsConfigVars in HCaptchaOutput, put the config vars
in a virtual config.json file in the RL module instead.

The VisualEditor module was already doing this, but was using a hook
where it didn't need to; simplify that too so we can delete the hook
file.

Using the same names for config settings on the server and on the client
makes it easier to write this code, but it also makes it easier to
search for all uses of a config setting.

Bug: T398204
Change-Id: I68c341d1b00f48fd0e99fef2ed5cfe1993aea295