hCaptcha: Fix mock for StatsFactory
We now have a unit test helper for StatsFactory that is recommended
for testing. Maybe the helper did not exist before, hence used to just
mock things.

This patch migrates the mock to the unit test helper because it caused
a patch to fail in core: Iac57f9c6794739b2ddaf3e8db661a852fe79efa5
that interacts with this code.

There is a snippet of the failures (comment section) in the patch
mentioned above since CI test logs will vanish after a couple of
days.

Change-Id: Ib057eb8aadd591acbfef0e9ec8df10b777655878