hCaptcha: Provide capabilities for failing over to alternate CAPTCHA type
Why:

- If MediaWiki is unable to load hCaptcha JS, or submissions to hCaptcha
  are failing, we should be able to failover to a different CAPTCHA type

What:

- Implement a `isAvailable` method on a new
  HCaptchaEnterpriseHealthChecker service
- Implement a "failover mode" that's activated when the number of
  SiteVerify calls exceeds a configured threshold, or if the hCaptcha
  JavaScript file is not accessible or fails integrity checks
  - The failover mode is set for 10 minutes, to allow time for manual
    intervention and also to not unnecessarily call endpoints that are failing
- Increment a counter for failed SiteVerify API calls (due to HTTP or
  JSON decoding errors), and cache this number for up to one minute
- Define configuration for the number of failed SiteVerify calls needed
  before entering a failover mode
  - Set the timeout to SiteVerify calls to 5 seconds
- Have the `isAvailable` method also check if the hCaptcha JavaScript
  API file is available, and that it meets the integrity checks, if
  configured
  - Use a 2 second timeout for calling the API URL, and retry once
    before implementing a failover mode.
- Keep track of how long isAvailable is taking to execute

Example hook implementation for LocalSettings.php:

```
$wgHooks['ConfirmEditCaptchaClass'][] = static function ( $action, &$className ) {
	if ( $className === 'HCaptcha' ) {
		/** @var HCaptchaEnterpriseHealthChecker $healthChecker */
		$healthChecker = MediaWikiServices::getInstance()->getService( 'HCaptchaEnterpriseHealthChecker' );
		if ( !$healthChecker->isAvailable() ) {
			LoggerFactory::getInstance( 'captcha' )->warning(
				'Falling back to FancyCaptcha due to hCaptcha health check failing.'
			);
			$className = 'FancyCaptcha';
		}
	}
};
```

Summary of workflow:

In the `ConfirmEditHooksGetInstance` hook, an implementer can call
`HCaptchaEnterpriseHealthChecker::isAvailable()`. This consults a
memcache key to check if "failover mode" is enabled. If failover mode is
set, then `::isAvailable()` returns false immediately, and the hook
implementer can use another captcha type. The TTL for failover mode is
10 minutes.

If we're not in failover mode, then we check two other memcache keys.

The first memcached key contains the count of SiteVerify API errors
(failed submissions to hCaptcha's API when a user submits a form).
If the count of these errors is over a defined threshold, then we
enter failover mode. The TTL for SiteVerify errors is 1 minute.

If the SiteVerify check is OK, then we check a memcache key which stores
the availability of the `secure-api.js` file. If the file is
downloadable, and if the integrity hash matches what we've configured,
then the status is good; if not, we enter failover mode. The
TTL for this data is 5 minutes.

Bug: T404204
Change-Id: I0870d90d4a740614a382cb6c2d0d36c3ffd6a2ed
(cherry picked from commit 53c82eb84f54da6e38c785d3074293828611a67b)