Skip QUnit tests requiring VisualEditor when it is not loaded
On CI, we would like to remove the recursive injection of extensions
dependencies. When ConfirmEdit is a dependency, we are required to add
VisualEditor to have tests to pass.

By making VisualEditor optional, on CI that saves us from having to add
VisualEditor as a dependency when it is otherwise not needed.

Bug: T389998
Change-Id: I8ded15ca57a996faa3f9726ee8780cb702091e59
(cherry picked from commit c0c2c6e9d999434f05f665d19b4e533e3cb5ea75)